Refinements to login page & installer
This needs to be tested!!!
Moved COWOA login page layout to it's OWN file. This eliminates the need
to have to turn on the split page login setting, and ensures that the
COWOA version of the login page shows ONLY when COWOA is turned on, and
that the "standard" login page works independently of the COWOA version.
Discovered that the new COWOA style of login page could be broken or
displayed incorrectly when the split login page setting was
inadvertently turned off or deleted..  Also turns off the display of the
"guest checkout" section of the login page when a site visitor clicks
the "login" link without items in the cart.

Re-worked the installer and removed the split page login update.